A Robust Method for Transcript Quantification with RNA-seq Data.
The advent of high throughput RNA-seq technology allows deep sampling of
the transcriptome, making it possible to characterize both the diversity
and the abundance of transcript isoforms. The software implemented a
general linear framework for accurate transcript quantification using a
set of new structural features: MultiSplices. Our software has several
desirable features: 1. It utilizes all the information implied in the
read alignment and alleviate the identifiability issues. 2. By solving
the linear system using LASSO, it can achieve the most accurate set of
dominantly expressed transcripts. 3. It is very efficient. For example,
the analysis of the human transcriptome can be finished in less than one
hour.